+++ .gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+.DS_Store/*
+
