+++ src/old/bot/README.md
@@ -1,0 +1,2 @@
+# NOTE
+this folder contain the code for previous version of 0xRythm with old discordjs version
