+++ src/old/bot/bot-config.ts
