+++ src/old/bot/bot.ts
