+++ tsconfig.json
@@ -1,17 +1,27 @@
 {
     "compilerOptions": {
-        "target": "es6",
-        "module": "commonjs",
-        "lib": ["ES2019"],
+        "removeComments": false,
+		"alwaysStrict": false,
+        "strict": false,
+        "target": "es2019",
+        "moduleResolution": "node",
+        "lib": ["ESNext", "DOM"],
+        "module": "CommonJS",
         "outDir": "dist",
         "baseUrl": "./",
-        "incremental": true,
-        "resolveJsonModule": true,
-        "sourceMap": true,
-        "experimentalDecorators": true,
         "emitDecoratorMetadata": true,
-        "esModuleInterop": true,
-        "allowSyntheticDefaultImports": true
+        "experimentalDecorators": true,
+        "allowSyntheticDefaultImports": true,
+        "inlineSources": true,
+        "rootDir": "src",
+        "sourceMap": true,
+        "pretty": true,
+        "allowJs": true,
+        "incremental": true,
+		"esModuleInterop": true,
+		"skipLibCheck": true,
+        "declaration": true,
+		"declarationMap": true
     },
     "include": ["src/**/*.ts"],
     "exclude": ["node_modules"]
